Remove invalid company search view filter fields
Also adds a test to validate that the filter fields exist in the serializer class for the view.
